# docs/swagger.yaml
basePath: /api
definitions:
  controller.BulkResolveRequest:
    properties:
      contexts:
        items:
          $ref: '#/definitions/controller.ResolveRequest'
        type: array
    type: object
  controller.Error:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  controller.ResolveRequest:
    properties:
      asOfDate:
        description: '"2006-01-02"'
        example: "2026-01-15"
        type: string
      caseType:
        example: judicial
        type: string
      client:
        example: Chase
        type: string
      investor:
        example: FNMA
        type: string
      state:
        example: FL
        type: string
    type: object
  controller.ResponseData:
    properties:
      data:
        description: Data is the response data.
    type: object
  controller.ResponseError:
    properties:
      errors:
        description: Errors is the response errors.
        items:
          $ref: '#/definitions/controller.Error'
        type: array
    type: object
...
  domain.CaseContext:
    properties:
      asOfDate:
        description: defaults to time.Now() if zero
        type: string
      caseType:
        type: string
      client:
        type: string
      investor:
        type: string
      state:
        type: string
    type: object
...
  domain.ResolvedConfig:
    properties:
      context:
        $ref: '#/definitions/domain.CaseContext'
      resolvedAt:
        type: string
      steps:
        additionalProperties:
          $ref: '#/definitions/domain.ResolvedStep'
        type: object
    type: object
  domain.ResolvedStep:
    additionalProperties:
      $ref: '#/definitions/domain.ResolvedTrait'
    type: object
  domain.ResolvedTrait:
    properties:
      overrideId:
        type: string
      source:
        description: '"default" or "override"'
        type: string
      value: {}
    type: object
  domain.Selector:
    properties:
      caseType:
        example: FC-Judicial
        type: string
      client:
        example: Chase
        type: string
      investor:
        example: FHA
        type: string
      state:
        example: FL
        type: string
    type: object
  domain.TraceCandidate:
    properties:
      effectiveDate:
        type: string
      outcome:
        description: '"SELECTED — ...", "SHADOWED — ..."'
        type: string
      overrideId:
        type: string
      selector:
        $ref: '#/definitions/domain.Selector'
      specificity:
        type: integer
      value: {}
    type: object
  domain.TraceSource:
    properties:
      effectiveDate:
        type: string
      overrideId:
        type: string
      selector:
        $ref: '#/definitions/domain.Selector'
      specificity:
        type: integer
    type: object
  domain.TraitTrace:
    properties:
      candidates:
        items:
          $ref: '#/definitions/domain.TraceCandidate'
        type: array
      resolvedFrom:
        allOf:
        - $ref: '#/definitions/domain.TraceSource'
        description: nil = resolved from default
      resolvedValue: {}
      step:
        type: string
      trait:
        type: string
    type: object
...
host: localhost:8082
info:
  contact: {}
  description: Foreclosure rules resolution engine — resolves step and trait values
    for a case context using a specificity-based override system.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Rules Resolution Service
  version: "1.0"
paths:
...
  /resolve:
    post:
      consumes:
      - application/json
      parameters:
      - description: Case context to resolve
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/controller.ResponseData'
            - properties:
                data:
                  $ref: '#/definitions/domain.ResolvedConfig'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ResponseError'
      summary: Resolve case context
      tags:
      - resolve
  /resolve/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Multiple case contexts to resolve
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.BulkResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ResponseError'
      summary: Bulk resolve case contexts
      tags:
      - resolve
  /resolve/explain:
    post:
      consumes:
      - application/json
      parameters:
      - description: Case context for trait trace
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controller.ResolveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TraitTrace'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/controller.ResponseError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/controller.ResponseError'
      summary: Explain trait resolution
      tags:
      - resolve
schemes:
- http
swagger: "2.0"